Split caliptra-registers into caliptra-registers-latest.
This is in preparation for a future caliptra-registers-1.0 crate.
