+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
   "drivers/test-fw/scripts/vector_gen",
   "fmc",
   "fmc/test-fw/test-rt",
+  "hw/latest/registers",
   "hw/verilated",
   "hw-model",
   "hw-model/test-fw",
@@ -120,6 +121,7 @@ caliptra-image-verify = { path = "image/verify", default-features = false }
 caliptra-kat = { path = "kat" }
 caliptra-lms-types = { path = "lms-types" }
 caliptra-registers = { path = "registers" }
+caliptra-registers-latest = { path = "hw/latest/registers" }
 caliptra-runtime = { path = "runtime", default-features = false }
 caliptra-systemrdl = { path = "systemrdl" }
 caliptra-test = { path = "test" }
