+++ hw/latest/registers/src/doe.rs
