+++ hw/latest/registers/src/entropy_src.rs
