+++ hw/latest/registers/src/hmac.rs
