+++ hw/latest/registers/src/soc_ifc.rs
