+++ hw/latest/registers/src/soc_ifc_trng.rs
