+++ hw/latest/registers/src/spi_host.rs
