+++ registers/update.sh
@@ -8,4 +8,4 @@ if [[ ! -f "../hw/latest/rtl/.git" ]]; then
     exit 1
 fi
 
-cargo run --manifest-path bin/generator/Cargo.toml -- ../hw/latest/rtl bin/extra-rdl/ src/
+cargo run --manifest-path bin/generator/Cargo.toml -- ../hw/latest/rtl bin/extra-rdl/ ../hw/latest/registers/src/
